requirements.txt
coverage
detect-secrets==1.0.3
myst-parser
pre-commit
pytest
python-dotenv
Sphinx
toml
ipykernel
google-api-core==1.34.1
google-api-python-client==1.8.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.75.0
google-cloud-artifact-registry==1.14.0
google-cloud-bigquery==3.25.0
google-cloud-bigquery-connection==1.17.0
google-cloud-bigquery-storage==2.27.0
google-cloud-core==2.4.1
google-cloud-datastore==1.15.5
google-cloud-functions==1.19.0
google-cloud-iam==2.17.0
google-cloud-jupyter-config==0.0.10
google-cloud-language==2.16.0
google-cloud-monitoring==2.24.0
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.14.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
plotly
matplotlib
numpy
pandas
db-dtypes
seaborn
ipython
geopandas
folium
mapclassify
ipywidgets
toolz
